Introducing Odyssey Support!
This update brings with it support for Odyssey and updates for Sortie. Boss chamber gadgets are now usable even if the player loaded or reloaded superwarp inside the boss chamber (or crashed); Before the player would simply receive a message saying you could not use the gadget in that scenario. Aminon warps now function properly, the player will be required to set the difficulty to the fight before warp will be allowed.