Warp targetting bug fixed.
This commit fixes the pesky bug where superwarp attempts to target a warp npc with the same name as the one it should be targetting bc the distance is showing as 0 or near 0 for that incorrect npc. This fix has superwarp find the true 3D distance. Solving that issue for good. Apologies I didn't solve this sooner, its just it had never happened to me until last night in temenos and then again in apollyon.