Added 'back' & 'random' commands; Added limbus map
Two new commands have been added to the temenos and apollyon maps to make getting the 5k and/or the matter augment item more likely by allowing the user more options for doing the climb in different orders. I have made a master map for temenos and apollyon. The user is now able to use one command prefix for temenos and apollyon. //li   or  /console li