Timed out waiting for response fix.
This update fixes the long wait time that occurs in certain scenarios, which these days seem more common than ever, where the user will stand for several seconds after inputting a warp command before the warp goes through. I have found the server response time from interacting with a warp npc to be at most 1.25 and in the most extreme example i've seen 1.3 seconds in sortie, so to be safe I added another .2 seconds on top of that for the first attempt (and hopefully only attempt) and if it is not received by that time, superwarp will immediately re-attempt initiating npc interaction which in 99.9% of cases will go through if it didn't on the first so I removed the log message for the first failed poke since the average time for the warp to go through on failed poke is very similar to if it hadn't failed at all. (About 2.5s)
   Secondly, I changed the default_packet_wait_timeout's  default setting from 5 to 3 and from what I can tell this is plenty of time for everything this variable is used for especially in the poke_npc() .
  This update is particularly important for Sortie basement exiting, but it is an all around quality improvement to the addon.